Switch recipe back to noarch: python
The existing recipe is much more complex and restrictive than it
needs to be.  This package doesn't compile anything so it
doesn't need dependencies in host and it doesn't need to pin
compatible versions in run.  This is causing trouble with
downstream packages that want to use more recent versions of
the packages than these restrictions allow.